test/e2e_tests/requirements.txt
requests
boto3
pyyaml
psutil